Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  1f9f2dced80b6114013cc43ecafd57aeb71cd499
Avoid forceHTTPS cookie flapping if core and CA are setting the same cookie

Ib7b0f77a mostly stopped CentralAuthSessionProvider from setting the
forceHTTPS cookie multiple times in one request, but if both core and CA
are setting the cookie with the same domain and path it will still
delete and reissue the cookie repeatedly. So detect that situation and
skip the "delete".

Bug: T124421
Change-Id: Ie74940d7bf4316acaaec037c6f0a3b92b96df380