Updated mediawiki/extensions Project: mediawiki/extensions/Echo  3f7436c95991792ea992031d679b3418246ce31b
Wrap CallbackFilterIterator in a condition check

Wrap the CallbackFilterIterator backport class in a conditional check
for PHP runtimes that include the class natively. This really should
only be needed for linting as the class is loaded via an autoloader
and thus should not be loaded if the runtime already has it
available.

Bug: T124828
Change-Id: I39d27385186d4693a8babdd2b818e6b4bc16255a