Updated mediawiki/extensions Project: mediawiki/extensions/Flow  f293f547708f9546b0a8e4cb0ba8bcc68d86deb2
Mentions are the primary notification

There was code in locateMentionedUsers to exclude users who already
got the other type of motification. Since we want mentions to be the
primary notification, we no longer need that.

In addition, we want those other messages to no longer go out if the
mention notification is being sent.

Bug: T125428
Depends-on: Ifa0e2d3283f57624af4c5ec264f9f66223508e83
Change-Id: I5cd29f475cadf86d12d67f8d3bd63407e761afc2