Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  587f3dff449f74eed544260c2112f41e57f5a665
Term box: Take changes in all languages into account

This issue was introduced with the optimizations done in Ia7cd83f.
There are not only the two event handlers that update the page title
and the description and aliases section, there is also one in the
toolbar that enables the save button. We forgot this.

Bug: T125813
Change-Id: I3e0746e576032f6ef10dde81a165c0cb13b67337