Updated mediawiki/extensions Project: mediawiki/extensions/BlueSpiceFoundation  6745851487e36c554eee12e94b731e01a90ed4d4
ExtJS/RL: wait for Ext.onReady

It seems that just wating for the CSS to be loaded is not sufficient.
Ext.onReady also waits for the framework to be fully initialized.

All extension code that creates Ext compnenents on page load should also
use the Ext.onReady wrapper.

Needs cherry-pick to master

Change-Id: I7f54a3d9735fe12fafb649f7f9f87667d9c7a432