Updated mediawiki/extensions Project: mediawiki/extensions/OAuth  008309f0f793000574fef749301cf6cd96489115
Handle error message in SpecialMWOAuthManageConsumers

DAO objects, when wrapped in MWOAuthDAOAccessControl, can return
a Message explaining that the user lacks permissions, instead of
the real value. SpecialMWOAuthManageConsumers expected the
'restrictions' field of MWOAuthConsumer to always be an MWRestrictions
object, which caused problems when trying to manage another user's
consumer without the mwoauthviewprivate right.

Bug: T125939
Change-Id: I0212234d10243cc39a2cb18eb8d1e0f620e3123b