Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  4e01f750131b93ddf3959358168eae0106ef6b88
Make SpecialSetLabelDescriptionAliases use ChangeOps.

The ChangeOps (plural!) class has special logic for batching validation
that ensures consistent checking of constraints that involve multiple
parts of an entity (such as checking the uniqueness of the combination of
label and description). SpecialSetLabelDescriptionAliases was not
using this, causing it to effectively bypass the uniqueness check.

Bug: T121395
Change-Id: I513c00708120feff12588855b884c5cd145aa4d8