Updated mediawiki/extensions Project: mediawiki/extensions/MP3MediaHandler  b618d6ef1d8ea8464b2988e110e60c0152d833d2
* Adding the MP3MediaHandler code, which was what this extension was originally intended to house.

This extension is not, as per the original credits, the MiniMP3 extension that was originally committed, but is actually a new extension inspired/based-on that one, which uses HTML5 <audio> tags for MP3 handling, with the possibility of a Flash fallback, if necessary.  After the various discussions in IRC, I thought it that my version of the code had been committed, but it looks like the MiniMP3 code was committed instead - possibly with the intention of my updating-it, but if so I didn't realise that was the case.

Anyway, this is the code that should have been in this repository from the start (well, an immediate revision 2, anyway), which is a more general MP3-renderer, using HTML5.

Change-Id: Ib831ac899494690f6d8d9d914cbe81e1c74481eb