Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  c3e027d2124cdcd5ce4ce85427d62eb02ca1541c
Do not delete the translation which was once published

If the translation in progress was once published, deleting it should move
it to published status.

There is no UX change. If the translator deletes it from 'In progress'
dashboard, it get deleted as usual, but clicking 'Published' tab
in dashboard will show it there.

Bug: T123902
Change-Id: Icfa258cb7f35805a8c93a6acb3860fc2c6002a7e