Updated mediawiki/extensions Project: mediawiki/extensions/PluggableAuth  5b0cc60de3f550808419430e347c397819fa0efa
Add hook to avoid sending redirects if needed

I needed this to allow an SSO-authorized crawler to crawl the site.

The issue I experienced may be the result of bugs in earlier versions of
MediaWiki and/or the result of bugs in the crawler that I have no
ability to fix.

Bug: T126742
Change-Id: Iebc46db741410186508522fe2e50952c6c2915dc