Updated mediawiki/extensions Project: mediawiki/extensions/MP3MediaHandler  82937e34fa187bd29f7f29e2482b61e1abf3bb79
I've fixed the Flash code that we were generating.  I was mistakenly outputting the mp3 name using the <param> argument, but if these are arguments to be passed into Flash they need to be handled in the <param name="FlashVars"> parameter.

I have therefore updated the code to work in this fashion, plus I have added a new config variable to allow additional FlashVars to be set ($wgFlashPlayerFlashVars).

Finally, I've added optional $wgFlashPlayerWidth and $wgFlashPlayerHeight settings to allow you to control the size of the SWF file.

Change-Id: I8fa563b47255f6a7efe57d11f04ea6e6f713b361