Updated mediawiki/extensions Project: mediawiki/extensions/UniversalLanguageSelector  63bc85649a5def7159383747141fede3f2c89b50
Ensure module 'jquery.tipsy' is loaded before using

Move the showTipsy calls into the callback.

Move the declaration of the functions hideTipsy and showTipsy before the
new call position.

Also change the order of the declaration of the functions hideTipsy and
showTipsy to avoid a forward reference in showTipsy. Spotted by JSLint.

Bug: T118507
Change-Id: I338a78cde249405459be14948b78cf59d715145a