Updated mediawiki/extensions Project: mediawiki/extensions/Flow  b563a6e2bfbc89a7a6df3442e63a9adb1d1cc062
Switch to BasicObjectMapper for TopicListEntry

Before, depending on what was loaded and ordering, either a full
(with last_updated) or regular (just what's in flow_topic_list)
version of TopicListEntry would be cached by CachingObjectMapper.

That meant that moving would fail on some boards due to topic
moderation.

The default ordering is now last updated, and moves
trigger the ordering issue and force a re-render that requires
an offset to avoid moderated topics.

I looked at an alternative:

Add a separate mapper for TopicListLastUpdatedStorage (even just
any separate mapper would work, but I was going to make the updated
one uses a tri-key PK including last updated).

This was somewhat feasible, but still turned out more invasive than
seemed justified (e.g. we would have needed a separate
ObjectManager just for this reason, and then a way to get that out
of ManagerGroup, and conditionally pass it to Pager...).

I doubt there are any actual significant performance issues with this
(fromStorageRow and toStorageRow are both simple for TopicListEntry).

At some point, we could consider changing how CachingObjectMapper
works to avoid this issue in general, or revisiting whether we need
this caching layer.

Bug: T126701
Change-Id: I43872e87b0cf181e1a351bb1e258645413d07fbb