Updated mediawiki/extensions Project: mediawiki/extensions/FlaggedRevs  df619a5d8a698e71f676f9d37f5f949c1a6f72c5
Apply the nomobile class

The ext.flaggedRevs.basic is not currently targeted at mobile and looking
at it, there is no need for it to be.

Unfortunately this means the banner in FlaggablePageview is visible in mobile
browsers taking up real estate.

This generic solution allows the hiding of this banner at mobile resolutions

Bug: T119344
Change-Id: I8c718394f870812ec946dffb2b89474757c64b08