Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaPageViewInfo  d79d6f7f355ff4a8699520232fd9d7185dc48a29
Use ProcessDialog to add close button and title

Use OO.ui.ProcessDialog instead of a plain OO.ui.Dialog so we can have a
close button and a title.

The title is the date range of when the data is from, so adjust the
backend code to make it available for the frontend.

Bug: T127280
Change-Id: I3becf2d6da45106fc26c4321dc8b64e7e508b510