Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  cceb75561f48bdc0ed2510ed58f2cfdfe7934930
Filter currencies by country

This commit changes the function signature from static and taking
no arguments to be an instance method that takes an array argument
to pass state information to an adaptor that may not yet have
staged data (as when constructing DonationData).

Only the Astropay adapter does any filtering with this data so far.

Bug: T124327
Change-Id: I46ed4f483c89bc8e1dbedd60c29bd570c5e47f8f