Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  9f51ec8d8bfb7ce5efcc4eb7c508f4d443f03154
Create cindy profile to be useable inside a Vagrant box

This profile is almost completely but not quite the same as the
mw-vagrant-host. It can be used to execute tests inside of Vagrant (or on
Cindy for that matter). I just added the *_api_* urls and changed the port.

As quite a few env variables are already set under /etc/profile.d, for this
profile to be effective, a ~vagrant/.profile is required, with the
following content:

PATH="/srv/Barry-the-Browser-Test-Bot:$PATH"

export MEDIAWIKI_ENVIRONMENT=mw-vagrant-host
unset BROWSER
unset MEDIAWIKI_API_URL
unset MEDIAWIKI_PASSWORD
unset MEDIAWIKI_URL
unset MEDIAWIKI_USER

This ensure that we are using configuration from environment.yml only and
seems to me to be easier to understand.

Note: I have no idea why we also have vars setup in /etc/profile.d, but as
this is shared with other users of mediawiki-vagrant, I don't want to
modify it without understanding more.

Note: this profile probably needs to be renamed "mw-vagrant-guest", because
it is the profile that *should* be used when running inside vagrant.

Change-Id: I9730b725d4a28fb01b49a24cd1987727fd6b5ba2