Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  d1e5505d1d4b46960aba454bc35b47d16485e180
Reduce memcache traffic for central session storage

Reduce read traffic by using a CachedBagOStuff, and write traffic by
skipping the write if nothing has changed and it's not relatively close
to expiring.

To handle the last bit, this moves the handling of the 'expiry' central
session variable into CentralAuthUtils::setCentralSession(). It seems
that it used to be used instead of the memcache expiry to expire
sessions, but 1896fc65 got rid of that and it has been unused ever
since so we may as well repurpose it.

Bug: T127772
Change-Id: I2cfcbf98f3f98416a8c2cd6664f07c602cf03259