Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  ecaa06862d6bf0809f32542901cf07857ed984c9
Rename repo folders to be PSR-4 compilant

I want to do this in multiple small steps: This patch does not touch
the namespace headers of these files. All it does is renaming folders.
That's all. This can not break anything (because of the way these
classes are currently loaded) and will make later PSR-4 related patches
much easier to review.

Change-Id: I3ec12ef17d54cff74ac261c2fef44f3ef32a4ba9