Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  17ce3142e56291f5beabac709fbf6ef4100ec13d
Hide categorization changes depending on user setting

Instead of hiding the categorization rows in the Watchlist completely,
without giving the user the possibility to show it, hide them when the
setting in the Preferences was set accordingly.

Currently there is no way to change this filter on the watchlist directly,
like on the desktop watchlist, however, other hide-options aren't here,
too.

Bug: T127723
Change-Id: I0844d1c27a33230627458bc7bd54e3f5c6638ae3