Updated mediawiki/extensions Project: mediawiki/extensions/BlueSpiceExtensions  5a4def2e3e2ddd59fa9f5699c697cc21d8300fa5
ExtendedEditBar: Enable only in certain contexts

ExtendedEditBar breaks when used in certain contexts, esp. in
combination with Extension:LiquidThreads. Now the toolbar
is disabled when the hook input is empty and on all actions not defined
in $bsgExtendedEditBarEnabledActions.

PatchSet 2: Implemented CR comments
PatchSet 3: As requested changed the logic to whitelist of actions

Change-Id: I90a5d9cb6b4edeb46a7a49d3d7ccf0ce69895b76