Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  c205530f28294c810fab617ad76842526b8825ff
Avoid using "new CentralAuthUser" since it avoids the cache

In the past, caching of CentralAuthUser was on the User object, so
anything that lacked a User object couldn't take advantage of caching.
Others needed to avoid the cache if they weren't sure the User object
wasn't going to be reset to anonymous or to a different name.

This was changed in I4199f56 to use a centralized cache keyed on the
username and independent of the specific User object. So we can now make
use of the caching in all the places where it was avoided before.

Maintenance callers and unit tests have not been updated in this patch.

Change-Id: Idc87313499c04b894308d4d14a578cc683b477f4