Updated mediawiki/extensions Project: mediawiki/extensions/WikibaseQualityExternalValidation  a3e04cf3881d11c1e6db78af8180e0b67a1e4cee
Add missing data-values/number dependency to composer.json

QuantityValueComparer uses the QuantityValue class, but it uses only
the absolute basic methods. The namespaces in the Number component never
changed. This means we are compatible with all versions of the Number
component we ever had. The planned 0.7.0 may include incompatible breaking
changes, that's why it must stop at 0.6.x.

Change-Id: Ie7a10c26fc9f04022757601d4cce2d10530b7506