Updated mediawiki/extensions Project: mediawiki/extensions/WikibaseQualityExternalValidation  142be947345b8de052d47e82b3ecfeeab1fdff2e
Add missing data-values/common dependency to composer.json

As far as I can see the only class used from this component is the
MonolingualTextValue class (and MultilingualTextValue, but we do not
support this property type anywhere).

Only very basic methods from MonolingualTextValue are used. The namespace
never changed. So we are compatible with all existing versions.

Also removes 2 unused use statements (unrelated to
the rest of the change)

Change-Id: I4a1b77550dd1f1c1e9830f106a3f42a290a76091