Updated mediawiki/extensions Project: mediawiki/extensions/Translate  d87eade09c6ce5886ff11a22db1a289947f0ea15
Do not prepend non-letter username with 'x'

Sanitizer::escapeId() prepends the result with a 'x' when the first character
is not a letter. This is necessary for the HTML attribute id.
This is not necessary here because the result gets prepended with
'tsb-request-' before using as HTML attribute id.
The option 'noninitial' deactivates this functionality.

Change-Id: I60af70b6682d7228a4435480b9856ded9b19bc36