Updated mediawiki/extensions Project: mediawiki/extensions/Echo  20f091982cb300bb6eca1e16b4c41877dfe91597
Don't list foreign notifications for both sections when not using groupbysection

API requests with notgroupbysection disabled but notsection set
to 'alert' would return foreign notification bundles for both
alerts and messages, separately.

The frontend always pasess notgroupbysection=1 in practice because
notmessageunreadfirst=1 doesn't work without it; but regardless of
that, the API module shouldn't return a foreign notification bundle
for messages when only alerts are requested.

Change-Id: I9fa6d1411a8c121591ce6dd75ee8052bcdcb478b