Updated mediawiki/extensions Project: mediawiki/extensions/WikibaseQualityConstraints  de9baba135af056bf93476a86eeef5d5c8210839
Add missing type hints everywhere

This fixes hundreds of code smell warnings, as reported by my PHPStorm.

When you want to review this, here is my suggestion:

1. Ignore all changes in comments. Even if there is a mistake (I'm pretty
sure there is not), comments are only documentation. A mistake in a comment
can not make the situation worse.

2. Only have a brief look on the changes in tests. When the tests succeed,
the changes I did are (almost certainly) fine.

3. Carefully review the changes in includes/…

Change-Id: Id0d8ca8e3f746cfd0c77f8824db918d495c635d4