Updated mediawiki/extensions Project: mediawiki/extensions/EventBus  23a0c2b92b8db79f83adb231c3b3b615b621e131
Added more protection against sending invalid events
to the eventlogging service.

Also, moved event construction out of the deferred task
for revision_visibility_set event, which greatly reduces
probability of a page being deleted simultaniously with
event creation. I suppose that was the reason for an NPE
that we've encountered in logs, but I'm not sure.

Bug: T128655
Change-Id: I1737de7ef1cb5478d9156b60f2f7daec3185c98c