Updated mediawiki/extensions Project: mediawiki/extensions/UrlShortener  51484e3100f08f1a4a7e221e8b6eb031550b10aa
Implement a read-only mode that prevents creation of new short URLs

We would like to get this extension deployed to Wikimedia sites so we
can prepare some surrounding infrastructure like Apache, Varnish, dumps,
etc., which should all be easier once the extension already exists.

This will allow us to install the extension but not allow any short
codes to be created in the database.

Change-Id: I648071bcb34259bb44b4d9a62f57d4c5bfdaac29