Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  289c2d9afa7793648fa58f4064526505d53bc2d6
Use PHP 5.5's …::class feature instead of strings in tests

This should be the final patch replacing all class names in strings
with the …::class keyword. This patch is limited to tests. The idea is:
When the tests succeed, all changes in this patch should be fine and
you do not need to do a super-critical character-by-character review.

Change-Id: I487b8180fee9c01385da7ce05e3f86628213bd88