Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  01bbaa23697ab89d245e841226b3d94b9b835dc5
Adding scripts to restart elasticsearch cluster.

There might be nicer ways to script / orchestrate things on our cluster (salt,
cluster SSH, ...) but in the meantime, those scripts are working, so they
should be available somewhere.

As discussed with ebernhardson, versionning them with the Cirrus extension
is not the worst place.

Bug: T129254
Change-Id: Icfc9fbe08261b8a7b73a6b03ff1f202d6c9bab17