Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  0744d72b8f7e060ec7cfce63bec1fbdee4db83f4
Minor improvements to script to restart elasticsearch cluster

Ugly but working: will try disabling / enabling replication until
it succeed. This is needed on eqiad, which has occasional timeouts.

This script it meant to be run under human supervision anyway, so having
not ideal error management (e.g. no max retries) is not blocking.

Change-Id: I7d294389c82fd313ad5a6f2d9f01da0dd2ad0307