Updated git submodules
Project: mediawiki/extensions/CentralAuth REL1_27 136eff6f6adcd70559feca457b6ebf7161306ef5

SECURITY: Check for valid but unusable user names

MediaWiki\Session\UserInfo::newFromName will throw an exception if
passed such a name, so it's no big deal when used with SessionManager.
But if used with older MediaWiki, if a name is usable on one wiki and
reserved on another in the same SUL grouping, you can log in on the
first and be logged in on the second when you shouldn't be able to.

Bug: T130384
Change-Id: I13cbe5294a03e4327a4edba56da3d99f5d3cb107