Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  cd08777bc1b22dea6e45d8b52d9eb95cbf964148
Disable problematic integration test

Until we can reproduce the failure locally (or debug the failure in the
integration environment) disable the "Checking that the preferred
language link has been created" scenario.

It has been enabled as a @smoke test which runs in a separate build
to allow us to re-evaluate this later but signal it is more important.

Bug: T130586
Change-Id: I81c2f45cfc8a60b644b1eefc4fbe8c8cfc769f74