Updated mediawiki/extensions Project: mediawiki/extensions/PageAssessments  86f6e3946b2ae87c615d0196ff98d2ac1fbdf1aa
Add PageAssessments API modules (projectpages and pageassessments)

This adds two API modules for PageAssessments: a query/prop module
to get assessment data for a pageset (ApiQueryPageAssessments), and
a query/list module (ApiQueryProjectPages, which can also be used
as a generator) to return pages associated with the given
WikiProject(s), with or without associated assessments.

This also moves WikiProject names to a separate table, so that the
data in the assessments table can be normalized (and the index can
be smaller).

Change-Id: I13b49573039b8b91739dc326455ee35d095b050f