Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  f520c3851fd6d92442ecbcb7dc2a55a1112817be
Do not bundle languages into subgroups

The new overlay has been incorrectly assuming that
languages prefixed with a code are related to the
language prefix. For example, it has assumed that
zh-min-nan was a variant of Chinese (zh) but in fact
it is a dialect
(https://en.wikipedia.org/wiki/Southern_Min)
This removes that assumption and gives all languages
an equal footing.

Bug: T130476
Change-Id: I8bb4579354fbefbaef9b181ee2f76d8fc94cca85