Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  dfee135619de46f032c93cdae29e230fed9b0a79
LanguageOverlay: send the correct position tapped value for EventLogging

When tapping on a language link, its position was being reported
incorrectly to the schema[1]. The problem arose when language links
were hidden but we were counting link wrappers, which were not hidden.
This patch correctly looks at language links and their classes to
determine which language link was clicked on.

[1] https://meta.wikimedia.org/wiki/Schema:MobileWebLanguageSwitcher

Bug: T130476
Change-Id: I79a45af0ac48a5133979bdbd8b64b2a8581b07fe