Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  b7232f996bd876d0450ad05db0310cf4cffd187f
Don't try to read a file to make thumbnails if there's none

This is apparently sometimes called if the upload is invalid (e.g.
unacceptable file extension) and this.file is not set up. This
probably shouldn't be happening, but let's just not throw exceptions.

Follow-up to 2b60bbbe1e19b5c6ba63e97e5749d34ceb10d0c6.

Change-Id: I30c6263640456ad7ec8e579a03c0a84668272fd4