Updated mediawiki/extensions Project: mediawiki/extensions/Comments  91f2c3268bd6dd24c0be08ccfb9ae29d33fb642c
Fix fatal caused by the lack of a proper Title when running some maintenance scripts (runJobs.php) etc.

I have no idea *why* the default is not sane, but that's what it is
without this setTitle() call here.

Tested on Brickipedia and seems to be working just fine; no more
Comments-related fatals when running runJobs.php.

Change-Id: I8aacdafa4d06268d9f20e4c1753a12875c50032b