Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaEvents  41c787ac36430e8fc073b9da044539e236b1ec84
Collect impact of proposed ResourceLoader feature-test in statsd

The sample factor of 1:1000 matches the ratio we currently use for collecting
Navigation Timing metrics.

This is a lightweight tracker that sends a beacon straight to statsv,
without the overhead of an EventLogging schema.

To get an idea of common user agents that don't pass, I plan to collect
samples from the kafka 'statsv' topic directly.

Bug: T128924
Change-Id: I672624e9fc3044c0dc024fa5097932c00f38d6db