Updated mediawiki/extensions Project: mediawiki/extensions/Math  d8e172f7138e60f3c89a6fbb395d093ca748ace7
Allow the number of concurrent connections to be configurable

The MultiHttpClient class allows the user to set the number of
concurrent connections it keeps open. This patch simply allows users of
the Math extension to set it when connecting to RESTBase.

Note that this patch is effectively a no-op, since MultiHttpClient's
default is (also) 50.

Bug: T132096
Change-Id: If9b6e61ffa23f8d22b3aab874058e809ff4c74b8