Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  e43fb56579835c9c6a8c7e2dd948bcc78cfa089b
Add script to run etsy/phan from in vagrant

Has the minor downside of breaking the usage of php5. The ppa used for
php7 forces upgrading to another set of packages from the ppa prefixed
with php5.6-*. The script takes care to install everything that should
be required to still utilizing mwscript and such, but it is an annoying
change that i would prefer not to have. Considered packaging php7
without those issues, but for the moment i think it's not required. If
we want to get this into the CI though we will have either find or make
compatible php7 packages.

Bug: T132625
Change-Id: Ibc09ec10b97ac8447875ec6036690db13b61fe5c