Updated mediawiki/extensions Project: mediawiki/extensions/CheckUser  91fb6c62cc1aece2744b4ba65b875a47f15d46c6
Change limit of 96 to 32 in CIDR calculator for IPv6

I don't know why 96 is the limit here but 96 is too small
to be useful and as far as I can tell, this does the
calculations correctly even for these ranges. Previously,
this was 64 but was changed in r47240
which doesn't state any reason.

Bug: T132893
Change-Id: Ic047ba1f1faed853759bc7313c8a022ca71e876c