Updated mediawiki/extensions Project: mediawiki/extensions/BetaFeatures  037cac99aaa1e69c13f66c4bc736ff02c0c2ddc3
Correction to test

The change in If4e0dfb6ee6674f0dace80a01850e2d0cbbdb47a
had a side effect of exposing a flaw in one of the BetaFeatures
tests. The default value of the NewHTMLCheckField was true before
but it was not rendering the checkbox as checked. Now this value is
honoured due to the call to HtmlFormField::getDefault which previously
was not called if the empty( $value->mParams['nodata'] ) was true.

Rather than change the HTML of the test, I've updated the setup
behaviour so that the checkbox is not checked initially.

Bug: T133163
Change-Id: Icb1e6ddc8f95da5e4b4de2916d292694c11ba731