Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  45ec156aff069db2a040954b3b6ff8067f11e11f
Use selection weights deterministically

The previous behavior was to roll a random number if more than one form
was available to handle the donor's context, and to use selection weight
as a probability to make some forms more or less likely to be chosen.

This patch makes a purely deterministic choice instead, using the highest
available weight.  We can add the A/B, random thing back later whenever we
want to use it intentionally.

We weight AstroPay more heavily so that it's always chosen in the countries
where it's supported.

Bug: T133181
Change-Id: I6f6959fea22b92c0b4e80d264b9aca881423442e