Updated mediawiki/extensions Project: mediawiki/extensions/PerformanceInspector  82810ed0b1bd9edbf4d7da411010da6a6a221cec
Focus on image size instead of page size

The old info module focused on page size even though we can't get the actual HTML size.
It's better for now to focus on only images (since that is what we can change as editors).

We now list all images and the size and display a warning if the image size is above the limit.

Change-Id: I141bec615e9d855e9142c6d51de1b58169e00c45