Updated mediawiki/extensions Project: mediawiki/extensions/CentralNotice  337ecafaef61d73a2d5a9c6b0cc513d45ae4b875
Mixed storage for impression diet and large banner limit

Adds a new mixed-storage system that allows fallback to cookies, and uses it
for impressionDiet and largeBannerLimit mixins.

For those mixins, we now only fall back to cookies for campaigns whose category
is in $wgCentralNoticeCategoriesUsingLegacy.

Cookies now use a format that mimics kvStore items. Legacy cookies will be
migrated.

Bug: T132639
Change-Id: Icd88f7bf9ad0d00fba6146f69691198fb4e9417f