Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  8562904636592011f38f12b426dfca098ddd83dc
configureHightingForSource expects an options field

When running without the experimental highlighter we end up raising an
undefined index warning because this options key only gets set in the
branch that deals with the experimental highlighter. This key is not
supported by the fvh, so don't provide it.

Add's some basic unit tests to ensure the config didn't change, beyond
stripping the options key, due to this adjustment. Also add's an
exception when trying to use highlight regex with fvh. This was already
unsupported, it just errored out in less obvious ways.

Change-Id: I26d6419341bec5d11a318bc50c07e18e57c7af1e