Updated mediawiki/extensions Project: mediawiki/extensions/Cite  b92a32e08cdd00c33ef0ef2f61a9ec41d6f8e6b0
Dialog: When reusing a reference, avoid popping up the group menu

The group picker menu will open momentarily when the input is changed, unless
it's disabled when the change happens. This was mostly visible when selecting
a reference from the "use an existing reference dialog", as a flash of the
menu in the top-left of the screen.

Bug: T125862
Change-Id: Ib6d3715f7d123545ba688bfbe850fe3c26ccb6e1