Updated mediawiki/extensions Project: mediawiki/extensions/BlueSpiceExtensions  40aed6cba6209c5349c28dfeca579ff6210610b8
Checklist: Fallback list options when list template is corrupt

When the list template indicated in checklist tag does not point to
an actual wiki page, display a "-" as the only list option. This resolves
an issue where the BlueSpice Visual Editor would not load when the list
template could not be resolved.

Change-Id: I521c791b994a57d2a9c5c7a70053304b481aaeae