Updated mediawiki/extensions Project: mediawiki/extensions/NavigationTiming  d261d407b434fa64a1bd4415b4687ce0fd408414
Verify chrome.loadTimes().firstPaintTime before reporting

We have entries in the Navigation Timing tables with values like
'1453951695461' for firstPaint. All have user-agents for Chrome or Chromium
version 20 or earlier. This is consistent with Chrome having a startLoadTime
value but no firstPaintTime value. To avoid reporting a bogus value in such
cases, verify that firstPaintTime > startLoadTime. If either measurement is
undefined, the comparison will return false.

Change-Id: Id5fd0f0efb2f148e861c1c30f38d4911b7efbd79